Refactor StateDashboard Component
